misc(@cubejs-backend/server): Improve code quality
